almost ready to use, but need makefile

+++ zypperzshcomp/test.sh
@@ -1,5 +1,4 @@
 cache="$HOME/.zzccache"
-
 [[ ! -d "$cache" ]] && mkdir "$cache"
 
 _cache_update() {
@@ -7,16 +6,42 @@ _cache_update() {
 	for d in /var/cache/zypp/raw/*; do
 		local pkgs=$d/suse/setup/descr/packages.gz;
 		local base=${d:t}
-		rm -f "$cache/$base"
+		local update=0
+		local date=0
+		
+		# figure when was this repo last cached
+		if [[ -e "$cache/$base" ]]; then
+			date=`stat -c "%Y" "$cache/$base"`
+		fi
+		
+		# figure out if it needs an update
 		if [[ -e "$pkgs" ]]; then
-			# using grep before sed really speeds things up
-			gunzip -c $pkgs | grep "^=Pkg: " | sed -n 's/^=Pkg:\s\s*\([^ ]*\).*/\1/p' >> "$cache/$base"
+			if [[ $date < `stat -c "%Y" "$pkgs"` ]]; then
+				update=1
+			fi
 		fi
 		for f in $d/repodata/*primary.xml.gz(N); do
-			gunzip -c $f | grep "<name>" | sed -n 's/.*<name>\(.*\)<\/name>.*/\1/p' >> "$cache/$base"
+			if [[ $date < `stat -c "%Y" "$f"` ]]; then
+				update=1
+			fi
 		done
 		
-		echo $base
+		#update the cache, if neccessary
+		if [[ $update == 1 ]]; then
+			rm -f "$cache/$base"
+			if [[ -e "$pkgs" ]]; then
+				if [[ $date < `stat -c "%Y" "$pkgs"` ]]; then
+					# using grep before sed really speeds things up
+					#gunzip -c $pkgs | grep "^=Pkg: " | sed -n 's/^=Pkg:\s\s*\([^ ]*\).*/\1/p' >> "$cache/$base"
+					# using my pkgnames thing is even much faster
+					zypperpkgnames "$pkgs" >> "$cache/$base"
+				fi
+			fi
+			for f in $d/repodata/*primary.xml.gz(N); do
+				#gunzip -c $f | grep "<name>" | sed -n 's/.*<name>\(.*\)<\/name>.*/\1/p' >> "$cache/$base"
+				zypperpkgnames "$f" >> "$cache/$base"
+			done
+		fi
 	done
 	for d in $cache/*(N); do
 		if [[ ! -e /var/cache/zypp/raw/${d:t} ]]; then
